python-service/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart>=0.0.6

# Database
sqlalchemy==2.0.20
alembic==1.11.1
psycopg2-binary==2.9.10

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.9.2

# Logging & HTTP
loguru==0.7.0
httpx==0.24.1
requests>=2.32.0,<3.0

# Testing
pytest==7.4.0

# AI/ML - UPDATED: Python 3.12 compatible versions with pre-built wheels
numpy==1.26.4
opencv-python-headless==4.10.0.84
openai>=2.7.1
sentence-transformers==3.0.1
torch>=2.0.0

# PDF Processing
PyMuPDF==1.26.5
PyPDF2==3.0.1
pdfplumber>=0.9.0
pytesseract==0.3.10
Pillow==10.1.0

# Excel Processing
openpyxl==3.1.2

# Fuzzy String Matching (FIXED for Python 3.12 compatibility)
rapidfuzz==3.5.2

# 🔄 Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# 🎥 Video Processing
openai-whisper==20231117
ffmpeg-python==0.2.0
yt-dlp==2023.12.30
youtube-transcript-api==0.6.1

# 📺 Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# 📦 Object Storage
minio==7.2.0